requirements.txt
streamlit>=1.32.0
groq>=0.9.0
numpy>=1.24.0
pypdf2>=3.0.0
pdfplumber>=0.10.0
python-pptx>=0.6.23
python-dotenv>=1.0.0
